requirements.txt
aiogram==3.25.0
aiosqlite==0.20.0
aiohttp==3.11.0
fastapi==0.135.0
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.0
itsdangerous==2.2.0
